cropRecommendationSystem/requirements.txt
absl-py==1.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
copulas==0.6.1
ctgan==0.5.1
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.1
deepecho==0.3.0.post1
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.9
djangorestframework==3.14.0
entrypoints==0.4
et-xmlfile==1.1.0
Faker==9.9.1
filelock==3.4.0
flatbuffers==22.10.26
gast==0.4.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.19.1
grpcio==1.34.1
h5py==3.1.0
hmmlearn==0.2.7
idna==2.10
importlib-metadata==4.6.0
importlib-resources==5.4.0
ipykernel==6.9.0
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.10.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==14.0.6
llvmlite==0.38.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mlxtend==0.19.0
nbclient==0.5.10
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numba==0.55.1
numpy==1.21.5
oauthlib==3.1.1
opencv-python==4.5.4.60
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.17.3
psutil==5.9.0
psycopg2==2.9.5
py-synthpop==0.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.14.1
Pygments==2.11.2
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pyts==0.12.0
pytz==2021.3
pywin32==303
pywinpty==2.0.2
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
randomgen==1.21.2
rdt==0.6.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.2
scikit-learn-extra==0.2.0
scipy==1.7.1
sdmetrics==0.4.1
sdv==0.14.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.15.0
sqlparse==0.4.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.6
virtualenv==20.10.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.2
wrapt==1.12.1
zipp==3.4.1
